requirements.txt
# requirements.txt

# Core packages
streamlit>=1.27
pandas>=2.0
numpy>=1.24

# NLP and Analysis
spacy>=3.7
sentence-transformers>=2.2.2
wordcloud>=1.9
nltk>=3.8

# Document processing
pdfplumber>=0.10.2
camelot-py>=0.11.0
opencv-python>=4.8.0
pytesseract>=0.3.10
docx2txt>=0.8
Pillow>=10.0.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Machine Learning
scikit-learn>=1.3.0

# Remember to run:
# python -m spacy download en_core_web_sm